requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.2
PyMySQL==1.1.1
python-dotenv==1.0.1
Werkzeug==3.0.3
ReportLab==4.2.2
openpyxl==3.1.5
python-barcode==0.15.1
Pillow==10.4.0
qrcode==7.4.2
cryptography==42.0.8
